Milestone: LId-driven cavity 2D works

+++ src/core/cell.hh
@@ -5,7 +5,7 @@
 
 
 template<typename T>
-__device__ T Cell<T>::computeEquilibriumPopulation(unsigned int l, const LBParams<T>*const params, T R, T U, T V) const {
+__device__ T Cell<T>::computeEquilibriumPopulation(unsigned int l, const LBParams<T>* const params, T R, T U, T V) const {
     int cix = params->LATTICE_VELOCITIES[l * params->D];
     int ciy = params->LATTICE_VELOCITIES[l * params->D+1];
     T cixcs2 = cix * cix - C_S_POW2;
@@ -19,67 +19,67 @@ __device__ T Cell<T>::computeEquilibriumPopulation(unsigned int l, const LBParam
 }
 
 template<typename T>
-__device__ T Cell<T>::getZeroMoment(const T*const population, const LBParams<T>*const params) const {
+__device__ T Cell<T>::getZeroMoment(const T* const population, const LBParams<T>* const params) const {
     T rho = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { rho += population[l]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { rho += population[l]; }
     return rho;
 }
 
 template<typename T>
-__device__ T Cell<T>::getFirstMomentX(const T*const population, const LBParams<T>*const params) const {
+__device__ T Cell<T>::getFirstMomentX(const T* const population, const LBParams<T>* const params) const {
     T m1x = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; }
     return m1x;
 }
 
 template<typename T>
-__device__ T Cell<T>::getFirstMomentY(const T*const population, const LBParams<T>*const params) const {
+__device__ T Cell<T>::getFirstMomentY(const T* const population, const LBParams<T>* const params) const {
     T m1y = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1y += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1y += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; }
     return m1y;
 }
 
 template<typename T>
-__device__ T Cell<T>::getVelocityX(const T*const population, const LBParams<T>*const params) const {
+__device__ T Cell<T>::getVelocityX(const T* const population, const LBParams<T>* const params) const {
     T m1x = 0.0;
     T rho = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; rho += population[l]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; rho += population[l]; }
     return m1x / rho;
 }
 
 template<typename T>
-__device__ T Cell<T>::getVelocityX(const T*const population, const LBParams<T>*const params, T R) const {
+__device__ T Cell<T>::getVelocityX(const T* const population, const LBParams<T>* const params, T R) const {
     T m1x = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D]; }
     return m1x / R;
 }
 
 template<typename T>
-__device__ T Cell<T>::getVelocityY(const T*const population, const LBParams<T>*const params) const {
+__device__ T Cell<T>::getVelocityY(const T* const population, const LBParams<T>* const params) const {
     T m1x = 0.0;
     T rho = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; rho += population[l]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; rho += population[l]; }
     return m1x / rho;
 }
 
 template<typename T>
-__device__ T Cell<T>::getVelocityY(const T*const population, const LBParams<T>*const params, T R) const {
+__device__ T Cell<T>::getVelocityY(const T* const population, const LBParams<T>* const params, T R) const {
     T m1x = 0.0;
-    for (size_t l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; }
+    for (unsigned int l = 0; l < params->Q; ++l) { m1x += population[l] * params->LATTICE_VELOCITIES[l * params->D + 1]; }
     return m1x / R;
 }
 
 template<typename T>
-__device__ void Cell<T>::getEquilibriumDistribution(const T*const population, T* eqDistr, const LBParams<T>*const params, T R, T U, T V) const {
-    for (size_t l = 0; l < params->Q; ++l)
+__device__ void Cell<T>::getEquilibriumDistribution(const T* const population, T* eqDistr, const LBParams<T>* const params, T R, T U, T V) const {
+    for (unsigned int l = 0; l < params->Q; ++l)
     {
 		eqDistr[l] = computeEquilibriumPopulation(l, params, R, U, V);
 	}
 }
 
 template<typename T>
-__device__ void Cell<T>::setEquilibriumDistribution(T* population, const LBParams<T>*const params, T R, T U, T V) const {
-    for (size_t l = 0; l < params->Q; ++l)
+__device__ void Cell<T>::setEquilibriumDistribution(T* population, const LBParams<T>* const params, T R, T U, T V) const {
+    for (unsigned int l = 0; l < params->Q; ++l)
     {
 		population[l] = computeEquilibriumPopulation(l, params, R, U, V);
 	}
@@ -87,7 +87,7 @@ __device__ void Cell<T>::setEquilibriumDistribution(T* population, const LBParam
 
 template<typename T>
 __device__ void Cell<T>::computePostCollisionDistributionBGK(T* population, const CollisionParamsBGK<T>*const params, T R, T U, T V) const {
-    for (size_t l = 0; l < params->Q; ++l) {
+    for (unsigned int l = 0; l < params->Q; ++l) {
         population[l] -= params->omegaShear * (population[l] - computeEquilibriumPopulation(l, params, R, U, V));// - Fext[l];	
 	}
 }
